support ruby 2.0
