+++ lib/kakine.rb
@@ -3,7 +3,7 @@ require 'kakine/cli'
 require 'kakine/cli/operation'
 require 'kakine/adapter'
 require 'kakine/resource'
-require 'kakine/resource/hash_sort'
+require 'kakine/hash_sort'
 require 'kakine/security_group'
 require 'kakine/security_group/diff_parser'
